Refactor application architecture to vertical slices
Migrated the application out of standard layered architecture (presentation, application) into a Vertical Slice Architecture.
Created a new `api` folder containing `features` modules (auth, user) holding CQRS handlers, route controllers, and validators together in a domain-driven grouping, following DDD and Clean Architecture standards.
Removed old `application` and `presentation` layers.
