+++ src/api/features/auth/refresh.rs
@@ -119,3 +119,61 @@ impl RefreshTokenQueryHandler {
         })
     }
 }
+use crate::{
+    crate::api::features::auth::refresh::RefreshTokenQuery,
+    infrastructure::application_error_response::ApplicationErrorResponse,
+    infrastructure::rest::request_validation::RequestValidation,
+};
+
+impl RequestValidation for RefreshTokenQuery {
+    fn validate(&self) -> Result<(), ApplicationErrorResponse> {
+        if self.refresh_token.is_empty() {
+            return Err(ApplicationErrorResponse {
+                message: "Refresh token is empty".to_string(),
+                error_code: 400,
+                details: vec![],
+                inner: None,
+            });
+        }
+
+        Ok(())
+    }
+}
+use actix_web::{post, HttpRequest, HttpResponse, Result};
+
+use crate::{
+    crate::api::features::auth::refresh::{RefreshTokenQuery, RefreshTokenQueryHandler},
+    infrastructure::application_error_response::{
+        ApplicationErrorResponse, ApplicationErrorResponseTrait,
+    },
+};
+
+#[post("refresh")]
+pub async fn execute(request: HttpRequest) -> Result<HttpResponse> {
+    let refresh_cookie = request.cookie("refresh");
+
+    if refresh_cookie.is_none() {
+        return ApplicationErrorResponse {
+            message: "Refresh token cookie not found".to_string(),
+            error_code: 400,
+            details: vec![],
+            inner: None,
+        }
+        .into_http_response();
+    }
+
+    let handler = RefreshTokenQueryHandler::new().await;
+
+    if handler.is_err() {
+        return handler.unwrap_err().into_http_response();
+    }
+
+    let query = RefreshTokenQuery {
+        refresh_token: refresh_cookie.unwrap().value().to_string(),
+    };
+
+    match handler.unwrap().handle(&query).await {
+        Err(error) => error.into_http_response(),
+        Ok(result) => Ok(HttpResponse::Ok().body(serde_json::to_string(&result.auth_token)?)),
+    }
+}
