+++ src/api/features/user/mod.rs
@@ -1,0 +1,10 @@
+pub mod create;
+pub mod get;
+
+use actix_web::web;
+
+pub fn user_routes(config: &mut web::ServiceConfig) {
+    config
+        .service(create::execute)
+        .service(get::execute);
+}
