--- src/application/auth/query/mod.rs
@@ -1,2 +1,0 @@
-pub mod auth_login_query;
-pub mod refresh_token_query;
