--- src/application/user/queries/mod.rs
@@ -1,1 +1,0 @@
-pub mod get_user_query;
