+++ src/lib.rs
@@ -1,5 +1,4 @@
-pub mod application;
+pub mod api;
 pub mod domain;
 pub mod infrastructure;
-pub mod presentation;
 pub mod routes;
