--- src/presentation/rest/auth/mod.rs
@@ -1,3 +1,0 @@
-pub mod auth_controller;
-pub mod refresh_controller;
-pub mod validate_controller;
