--- src/presentation/rest/mod.rs
@@ -1,4 +1,0 @@
-pub mod auth;
-pub mod default_response;
-pub mod user;
-pub mod validations;
