--- src/presentation/rest/user/mod.rs
@@ -1,2 +1,0 @@
-pub mod create_user_controller;
-pub mod get_user_controller;
