--- src/presentation/rest/validations/auth/mod.rs
@@ -1,2 +1,0 @@
-pub mod auth_login_query_validations;
-pub mod refresh_token_query_validations;
