--- src/presentation/rest/validations/mod.rs
@@ -1,2 +1,0 @@
-pub mod auth;
-pub mod user;
