--- src/presentation/rest/validations/user/commands/mod.rs
@@ -1,1 +1,0 @@
-pub mod create_user_command_validations;
